passes param to initiate task startup
